refactor: connectivity plot files
