+++ docs/api/_toctree/VisualizationAPI/miv.visualization.connectivity.plot_connectivity.rst
@@ -1,0 +1,6 @@
+﻿miv.visualization.connectivity.plot\_connectivity
+=================================================
+
+.. currentmodule:: miv.visualization.connectivity
+
+.. autofunction:: plot_connectivity
