+++ docs/api/_toctree/VisualizationAPI/miv.visualization.connectivity.plot_connectivity_interactive.rst
@@ -1,0 +1,6 @@
+﻿miv.visualization.connectivity.plot\_connectivity\_interactive
+==============================================================
+
+.. currentmodule:: miv.visualization.connectivity
+
+.. autofunction:: plot_connectivity_interactive
