--- docs/api/_toctree/VisualizationAPI/miv.visualization.connectivity_plot.plot_connectivity.rst
@@ -1,6 +1,0 @@
-﻿miv.visualization.connectivity\_plot.plot\_connectivity
-=======================================================
-
-.. currentmodule:: miv.visualization.connectivity_plot
-
-.. autofunction:: plot_connectivity
