+++ miv/visualization/__init__.py
@@ -1,5 +1,5 @@
 from miv.visualization.causality import *
-from miv.visualization.connectivity_plot import *
+from miv.visualization.connectivity import *
 from miv.visualization.event import *
 from miv.visualization.fft_domain import *
 from miv.visualization.raw_signal import *
